requirements.txt
Django==1.7.1
Pillow==2.6.1
PyJWT==0.3.0
django-grappelli==2.6.2
django-redis-cache==0.13.0
django-simple-captcha==0.4.2
djangorestframework==2.4.3
gunicorn==19.1.1
hiredis==0.1.5
oauthlib==0.6.3
python-openid==2.2.5
python-social-auth==0.2.1
redis==2.10.3
requests==2.4.3
requests-oauthlib==0.4.2
six==1.8.0
sorl-thumbnail==11.12.1b
wsgiref==0.1.2
